+++ modules/apps/object/object-api/src/main/resources/com/liferay/object/entry/util/packageinfo
@@ -1,1 +1,1 @@
-version 2.4.0
+version 2.5.0
